Fix browser REPL binding identity and var persistence
